option to skip docker import

+++ docker-create-debian-image.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 set -eu -o pipefail
 
+declare -r SKIP_DOCKER_IMPORT=${SKIP_DOCKER_IMPORT:-"false"}
+
 declare -r GREADLINK=$(which greadlink)
 declare -r READLINK=${GREADLINK:-"$(which readlink)"}
 
@@ -17,10 +19,14 @@ if [ -e "${DEBOOTSTAP_TAR}" ] ; then
 fi
 
 echo "debootstrab in a docker container"
-docker run -it --privileged --rm -v "${BASE_DIR}":"${BASE_DIR_DOCKER}" debian:stretch-slim bash -c "\"${DOCKER_SCRIPT}\" \"${DEBOOTSTAP_DIR_NAME}\" ; mv \"${DEBOOTSTAP_DIR_NAME}.tar\" \"${BASE_DIR_DOCKER}\""
+docker run -it --privileged --rm -v "${BASE_DIR}":"${BASE_DIR_DOCKER}" debian:stretch-slim \
+    bash -c "\"${DOCKER_SCRIPT}\" \"${DEBOOTSTAP_DIR_NAME}\" ; mv \"${DEBOOTSTAP_DIR_NAME}.tar\" \"${BASE_DIR_DOCKER}\""
 
-echo "create docker image"
-docker import "${DEBOOTSTAP_TAR}" debootstrap-stretch
-rm "${DEBOOTSTAP_TAR}"
+
+if [[ "${SKIP_DOCKER_IMPORT}" == "false" ]] ; then
+    echo "create docker image"
+    docker import "${DEBOOTSTAP_TAR}" debootstrap-stretch
+    rm "${DEBOOTSTAP_TAR}"
+fi
 
 echo "done"
